backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
SQLAlchemy==2.0.28
Werkzeug==3.0.1
cryptography==42.0.5